refactor: migrate Navbar component to TypeScript and update functionality
