--- frontend/src/components/navbar/index.js
@@ -1,3 +1,0 @@
-import { Navbar } from "./Navbar";
-
-export { Navbar };
